update WalkerProfile into components
