+++ client/src/App.js
@@ -69,7 +69,6 @@ function App() {
     const result = Auth.getProfileType();
     console.log(result);
 
-
     const [currentHeaderLink, setHeaderCurrentLink] = useState(headerLinks[0])
     const [currentWalkerLink, setWalkerLink] = useState(walkerLinks[0]);
 	const [currentOwnerLink, setOwnerLink] = useState(ownerLinks[0])
